+++ src/losant/contains-all.test.js
@@ -1,0 +1,17 @@
+const { testCases } = require('../../test/utils');
+const containsAll = require('./contains-all');
+
+testCases(containsAll, [
+  [
+    [[1, 3, 2], [1, 2, 3, 4, 5]],
+    true
+  ],
+  [
+    [[1, 2, 3], [1, 2]],
+    false
+  ],
+  [
+    [[3, 2, 1], [1, 2, 3]],
+    true
+  ]
+]);