+++ src/losant/key-by-with.test.js
@@ -2,7 +2,7 @@ const {
   pipe, unary, map, prop, always, tryCatch
 } = require('ramda');
 
-const testHarnessUnary = require('./test-harness-unary');
+const { testHarnessUnary } = require('../../test/utils');
 const keyByWith = require('./key-by-with');
 
 const testHarnessAdapter = (obj) => {