+++ src/losant/test-harness.js
@@ -1,12 +1,15 @@
-const {
-  curry, compose, evolve, map
-} = require('ramda');
-const format = require('./format');
-const stringify = require('./stringify');
-
 // TODO: Write tests for this?
 // TODO: Write README entry for this?
+// TODO: https://github.com/Losant/losant-utils/issues/40
 // NOTE: See `string-to-boolean` tests for example usage.
+
+const curry = require('ramda/src/curry');
+const compose = require('ramda/src/compose');
+const evolve = require('ramda/src/evolve');
+const map = require('ramda/src/map');
+
+const format = require('./format');
+const stringify = require('./stringify');
 
 const testCaseToString = compose(
   format('[{args}] ==> {expected}'),
@@ -16,8 +19,8 @@ const testCaseToString = compose(
   }),
 );
 
-// testHarness :: Function -> [*] -> *
-const testHarness = curry((fn, args, expected) => {
+// testHarness :: Expect -> Function -> [*] -> *
+const testHarness = curry((expect, fn, args, expected) => {
   it(
     testCaseToString({ args, expected }),
     () => expect(fn(...args)).toEqual(expected),